refactor: classroom people components file refactor for reusability
