+++ modules/classroom/components/people/peopleDetail/PeopleDetail.tsx
