+++ modules/classroom/components/people/peopleDetail/index.ts
