+++ modules/classroom/components/people/selfUserDetail/SelfUserDetail.tsx
