+++ modules/classroom/components/people/selfUserDetail/index.ts
