+++ modules/classroom/components/people/teacherDetail/TeacherDetail.tsx
