+++ modules/classroom/components/people/teacherDetail/index.ts
