+++ modules/classroom/containers/classroomContainer/ClassroomContainer.tsx
@@ -14,7 +14,7 @@ export default function ClassroomContainer() {
   const { id } = router.query;
   const classroomId = id as string;
 
-  const [getOneClassroom, { isLoading, isError, data: classroomData, error }] =
+  const [getOneClassroom, { isLoading, isError, data: classroomData }] =
     useLazyGetOneClassroomQuery();
 
   useEffect(() => {
@@ -27,14 +27,12 @@ export default function ClassroomContainer() {
 
   if (isLoading) return <LoadingComponent visible />;
 
-  if (isError) {
-    console.error(error);
+  if (isError)
     return (
       <div className="text-red-500">Failed to load classroom data. Please try again later.</div>
     );
-  }
+
   if (classroomData) {
-    console.log(classroomData.participants?.length);
     return (
       <div className="mx-8">
         <Tabs defaultValue="stream">
