+++ modules/classroom/containers/peopleContainer/PeopleContainer.tsx
@@ -7,7 +7,7 @@ import { EUserRole } from "@/shared/redux/rtk-apis/auth/auth.types";
 import type { IClassroomResponse } from "@/shared/redux/rtk-apis/classrooms/classrooms.types";
 import { TRootState } from "@/shared/redux/store";
 
-import TeacherDetail from "../../components/teacherDetail";
+import TeacherDetail from "../../components/people/teacherDetail";
 
 export default function PeopleContainer({
   owner,
